[feature/storage] Move all drivers to extension
