--- phpBB/phpbb/storage/driver/local.php
@@ -1,20 +1,0 @@
-<?php
-namespace phpbb\storage\driver;
-
-class local extends \phpbb\storage\driver
-{
-	public function __construct()
-	{
-		$this->filesystem = new \phpbb\storage\adapter\local();
-	}
-
-	public static function get_name()
-	{
-		return 'LOCAL';
-	}
-
-	public static function get_params()
-	{
-		return array();
-	}
-}
